GUACAMOLE-2185: Port numbers should be validated by guacd.

+++ src/protocols/vnc/settings.h
@@ -47,7 +47,7 @@ typedef struct guac_vnc_settings {
     /**
      * The port of the VNC server (or repeater) to connect to.
      */
-    int port;
+    unsigned short port;
 
     /**
      * The username given in the arguments.
@@ -112,7 +112,7 @@ typedef struct guac_vnc_settings {
     /**
      * The VNC port to connect to, if using a repeater.
      */
-    int dest_port;
+    unsigned short dest_port;
 #endif
 
 #ifdef ENABLE_VNC_LISTEN
